backend/requirements/dev.txt
-r base.txt
pytest>=8.0
pytest-django>=4.8
responses>=0.25
factory-boy>=3.3
ruff>=0.5
django-debug-toolbar>=4.4
